requirements.txt
openai
pinecone-client
python-dotenv
Langchain==0.0.225
pypdf